requirements_conda.txt
numpy
pytorch
tqdm
matplotlib
seaborn
scipy
scikit-learn
xarray
dask
netCDF4
numba
cartopy
